frontend/llm: reuse project context for ai gate
